refactor(admin): split problems route panels
Why:
- keep the Quantercise problems page as a route shell only
- separate query orchestration from filter, list, row, and pagination UI
- continue the admin cleanup train without changing route behavior

Verify:
- pnpm exec prettier --check apps/admin/src/app/quantercise/problems/page.tsx apps/admin/src/app/quantercise/problems/problems-content.tsx apps/admin/src/app/quantercise/problems/problems-sections.tsx
- git diff --check -- apps/admin/src/app/quantercise/problems/page.tsx apps/admin/src/app/quantercise/problems/problems-content.tsx apps/admin/src/app/quantercise/problems/problems-sections.tsx
- pnpm --filter @anipotts/admin typecheck
- pnpm --filter @anipotts/admin lint
- pnpm --filter @anipotts/admin build
- pnpm validate
